softirq: Fix lock imbalance in local_bh{disable,enable} due to preempt
By checking if the current context is preemptible to determine whether the
softirq lock should be locked/unlocked, it is therefore possible for both
the softirq lock to not be unlocked and for the softirq lock to not be
locked when it should be.

Example of softirq lock not getting unlocked:
	local_bh_disable(); <-- preemptible(), lock is acquired
	preempt_disable();
	local_bh_enable();  <-- !preemptible(), lock isn't released
	preempt_enable();

Example of softirq lock not being locked when it should be:
	preempt_disable();
	local_bh_disable(); <-- !preemptible(), lock isn't acquired
	preempt_enable();
	// !bh-safe section <-- preemptible(), non bh-safe section
	local_bh_enable();

Fix both issues by acquiring/releasing the softirq lock unconditionally.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>