bbdpl: Optimize IRQ for RT
The IRQ handler here masks the IRQ line and kicks off a worker which
unmasks it later. On RT this IRQ is force-threaded, so optimize it on RT by
making it non-threaded and eliminating the unnecessary sleepable irq_lock.

Note that gpio_to_desc() uses a sleepable spin lock so it needs to be moved
out from the IRQ handler.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>