vh: sched: Teach EAS about dm-constraints for CL0
Tensor's L3 cache shares a clock plane with CL0, which causes the L3$'s
performance to be dependent upon CL0's frequency. Because L3$ performance
affects the other two CPU clusters (CL1 and CL2), a constraint table is
used to set a frequency floor on CL0 depending on the frequency selected
for CL1 or CL2.

This is an unfortunate hardware limitation that is made more unfortunate by
the fact that EAS is unaware of the constraint system. Since there's no way
to communicate inter-cluster frequency dependencies in the energy model,
EAS doesn't know that dm-constraints exists and therefore makes suboptimal
energy decisions.

Specifically, EAS may think that waking a task on CL1 or CL2 may be
superior energy-wise even if it means that CL1 or CL2 need to run faster.
However, EAS doesn't know that making CL1 or CL2 run faster can have the
side effect of also making CL0 run faster, which in turn increases energy.

Teach EAS about dm-constraints for CL0 so that it can make better energy
decisions accordingly, knowing that the frequency of CL1 and CL2 can affect
the frequency of CL0 and thus alter predicted energy.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>