vh: sched: Run sugov IRQ worker from hard IRQ context
On RT, soft IRQ workers can be batched and delayed, and are then run from
an irq_work kthread. This adds a lot of latency to sugov's attempt to wake
its kthreads for CPU frequency switches, hurting performance.

Run the IRQ worker from hard IRQ context to force it to trigger an IPI
immediately for running the IRQ worker as soon as possible.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>