locking/rtmutex: Avoid implicit scheduling in lock slowpaths
When a lock is taken and it is determined that a waiter must sleep, it is
possible for the waiter to implicitly schedule before the explicit schedule
point. This hurts throughput and doesn't improve latency in any meaningful
way since there's already an explicit scheduling point.

Avoid implicit scheduling in lock slowpaths by disabling preempt prior to
the raw spin-lock unlocks which precede the explicit scheduling point.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>