eh: Decouple decompression command sets from specific CPUs
When pages are decompressed through eh, preemption is disabled via
get_cpu() and each decompression command set in eh is bound to a specific
CPU.

This is very much suboptimal for several reasons:
1. Preemption may be disabled for up to EH_POLL_DELAY_MS milliseconds.
2. Available decompression command sets may not be used simply due to a
   matter of CPU affinity.
3. CPU hotplugging results in decompression command sets belonging to
   offline CPUs never getting utilized.
4. Decompression can get bottlenecked when a single command set bound to a
   CPU is overutilized.

As it turns out, there's no inherent limitation in eh hardware which
requires that a decompression command set be bound to a specific CPU.
Preemption is only disabled because a single decompression command set must
not be used concurrently, and because it protects the per-CPU bounce
buffers.

Rather than binding decompression command sets to specific CPUs, decouple
them so that they can be flexibly used across all CPUs. This is done by
using a counting semaphore along with a bitmap in order to dynamically
utilize available decompression command sets without needing to disable
preemption.

This not only improves system latency by eliminating the get_cpu(), but
also improves eh's decompression performance by fully utilizing the
hardware to balance decompression across all command sets.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>